% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exasol.R
\docType{data}
\name{ALLOWED_UDF_IN_TYPES}
\alias{ALLOWED_UDF_IN_TYPES}
\title{All input types of UDF scripts}
\format{An object of class \code{character} of length 2.}
\usage{
ALLOWED_UDF_IN_TYPES
}
\description{
All input types of UDF scripts
}
\keyword{datasets}
