% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exasol.R
\docType{data}
\name{C_asyncEnableTracing}
\alias{C_asyncEnableTracing}
\title{Activates debug logs in the C layer.}
\format{
An object of class \code{CallRoutine} (inherits from \code{NativeSymbolInfo}) of length 4.
}
\usage{
C_asyncEnableTracing
}
\description{
Activates debug logs in the C layer.
}
\keyword{internal}
