% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exasol.R
\docType{data}
\name{C_asyncRODBCIOStart}
\alias{C_asyncRODBCIOStart}
\title{TODO comment}
\format{An object of class \code{CallRoutine} (inherits from \code{NativeSymbolInfo}) of length 4.}
\usage{
C_asyncRODBCIOStart
}
\description{
TODO comment
}
\keyword{datasets}

