% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exasol.R
\docType{data}
\name{C_asyncRODBCProxyPort}
\alias{C_asyncRODBCProxyPort}
\title{Returns the Database proxy port. Check the developer_guide for more information.}
\format{
An object of class \code{CallRoutine} (inherits from \code{NativeSymbolInfo}) of length 4.
}
\usage{
C_asyncRODBCProxyPort
}
\description{
Returns the Database proxy port. Check the developer_guide for more information.
}
\keyword{internal}
