% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exasol.R
\docType{data}
\name{C_asyncRODBCQueryStart}
\alias{C_asyncRODBCQueryStart}
\title{Starts an asynchronous query using the highspeed data channel. Check the developer_guide for more information.}
\format{
An object of class \code{CallRoutine} (inherits from \code{NativeSymbolInfo}) of length 4.
}
\usage{
C_asyncRODBCQueryStart
}
\arguments{
\item{chan}{An open Exasol connection}

\item{query}{The SQL query describing the Export/Import}

\item{protocol}{Http/Https}

\item{writer}{Indicating if it's a Export or Import. If 0, it indicates an import (import from DB into R) query, otherwise an export (export from R to the DB) query.}
}
\description{
Starts an asynchronous query using the highspeed data channel. Check the developer_guide for more information.
}
\keyword{internal}
