% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exasol.R
\docType{data}
\name{C_run_testthat_tests}
\alias{C_run_testthat_tests}
\title{Run C++ tests.}
\format{
An object of class \code{CallRoutine} (inherits from \code{NativeSymbolInfo}) of length 4.
}
\usage{
C_run_testthat_tests
}
\description{
Run C++ tests.
}
\keyword{internal}
