% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXADBI.R
\name{dbApply}
\alias{dbApply}
\title{Applies an R function to a result set.
The R code is transfered and deployed in an EXASOL database and executed massively parallel.}
\usage{
dbApply(res, fun, ...)
}
\arguments{
\item{res}{An EXAResult set.}

\item{fun}{A function to be applied to the result set.}

\item{simplify}{A logical indicating whether the result set may be transfered to the client and
stored in a data.frame. Default: FALSE.}
}
\value{
An EXAResult object relating to the result set in the EXASOL database, or if SIMPLIFY=TRUE
     a data.frame containing the result set.
}
\description{
Applies an R function to a result set.
The R code is transfered and deployed in an EXASOL database and executed massively parallel.
}
