% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXADBI-transaction.R
\name{dbBegin,EXAConnection-method}
\alias{dbBegin,EXAConnection-method}
\title{dbBegin}
\usage{
\S4method{dbBegin}{EXAConnection}(conn)
}
\arguments{
\item{conn}{An EXAConnection object}
}
\value{
A logical indicating success.
}
\description{
dbBegin
}
\section{Methods (by class)}{
\itemize{
\item \code{dbBegin(EXAConnection)}: Starts a DB transaction. In EXASOL, it disables autocommit.

}}
\seealso{
\code{\link[DBI:dbBegin]{DBI::dbBegin()}}

Other EXAConnection related objects:
\code{\link{EXAConnection-class}},
\code{\link{dbCommit,EXAConnection-method}},
\code{\link{dbConnect,EXADriver-method}},
\code{\link[=dbCurrentSchema]{dbCurrentSchema()}},
\code{\link[=dbEnd]{dbEnd()}},
\code{\link{dbExistsTable,EXAConnection,character-method}},
\code{\link{dbGetQuery,EXAConnection,character-method}},
\code{\link{dbListFields,EXAConnection,character-method}},
\code{\link{dbListTables,EXAConnection-method}},
\code{\link{dbReadTable,EXAConnection,character-method}},
\code{\link{dbRemoveTable,EXAConnection,character-method}},
\code{\link{dbRollback,EXAConnection-method}},
\code{\link{dbSendQuery,EXAConnection,character-method}},
\code{\link{dbWriteTable,EXAConnection,character,data.frame-method}}

Other transaction management functions:
\code{\link{dbCommit,EXAConnection-method}},
\code{\link[=dbEnd]{dbEnd()}},
\code{\link{dbRollback,EXAConnection-method}}
}
\author{
EXASOL AG <opensource@exasol.com>
}
\concept{EXAConnection related objects}
\concept{transaction management functions}
