% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXADBI-query.R
\name{dbClearResult,EXAResult-method}
\alias{dbClearResult,EXAResult-method}
\title{dbClearResult}
\usage{
\S4method{dbClearResult}{EXAResult}(res, ...)
}
\arguments{
\item{res}{An EXAResult object.}

\item{...}{Further arguments to passed to res$close(). This may be 'commit=TRUE' (not advisable).}
}
\value{
A logical indicating success.
}
\description{
dbClearResult
}
\section{Methods (by class)}{
\itemize{
\item \code{dbClearResult(EXAResult)}: Frees all resources associated with an \code{EXAResult}.

}}
\seealso{
\code{\link[DBI:dbClearResult]{DBI::dbClearResult()}}

Other EXAResult related objects:
\code{\link{EXAResult-class}},
\code{\link{dbListFields,EXAConnection,character-method}},
\code{\link{fetch,EXAResult,numeric-method}}

Other DQL functions:
\code{\link{dbExistsTable,EXAConnection,character-method}},
\code{\link{dbGetQuery,EXAConnection,character-method}},
\code{\link{dbListFields,EXAConnection,character-method}},
\code{\link{dbListTables,EXAConnection-method}},
\code{\link{dbReadTable,EXAConnection,character-method}},
\code{\link{dbSendQuery,EXAConnection,character-method}},
\code{\link{fetch,EXAResult,numeric-method}}
}
\author{
EXASOL AG <opensource@exasol.com>
}
\concept{DQL functions}
\concept{EXAResult related objects}
