% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXADBI-transaction.R
\name{dbCommit,EXAConnection-method}
\alias{dbCommit,EXAConnection-method}
\title{dbCommit}
\usage{
\S4method{dbCommit}{EXAConnection}(conn, silent = FALSE)
}
\arguments{
\item{conn}{An EXAConnection object}

\item{silent}{A logical, if FALSE no log messages will be exhibited.}
}
\value{
A logical indicating success.
}
\description{
dbCommit
}
\section{Methods (by class)}{
\itemize{
\item \code{dbCommit(EXAConnection)}: Sends a commit.

}}
\seealso{
\code{\link[DBI:dbCommit]{DBI::dbCommit()}}

Other EXAConnection related objects:
\code{\link{EXAConnection-class}},
\code{\link{dbBegin,EXAConnection-method}},
\code{\link{dbConnect,EXADriver-method}},
\code{\link[=dbCurrentSchema]{dbCurrentSchema()}},
\code{\link[=dbEnd]{dbEnd()}},
\code{\link{dbExistsTable,EXAConnection,character-method}},
\code{\link{dbGetQuery,EXAConnection,character-method}},
\code{\link{dbListFields,EXAConnection,character-method}},
\code{\link{dbListTables,EXAConnection-method}},
\code{\link{dbReadTable,EXAConnection,character-method}},
\code{\link{dbRemoveTable,EXAConnection,character-method}},
\code{\link{dbRollback,EXAConnection-method}},
\code{\link{dbSendQuery,EXAConnection,character-method}},
\code{\link{dbWriteTable,EXAConnection,character,data.frame-method}}

Other transaction management functions:
\code{\link{dbBegin,EXAConnection-method}},
\code{\link[=dbEnd]{dbEnd()}},
\code{\link{dbRollback,EXAConnection-method}}
}
\author{
EXASOL AG <opensource@exasol.com>
}
\concept{EXAConnection related objects}
\concept{transaction management functions}
