% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXADBI-connection.R
\name{dbConnect,EXADriver-method}
\alias{dbConnect,EXADriver-method}
\alias{dbConnect,character-method}
\alias{dbConnect,EXAConnection-method}
\title{dbConnect}
\usage{
\S4method{dbConnect}{EXADriver}(
  drv,
  exahost = "",
  uid = "",
  pwd = "",
  schema = "SYS",
  encryption = "Y",
  sslcertificate = "",
  autocommit = "Y",
  querytimeout = "0",
  ...
)

\S4method{dbConnect}{character}(drv, ...)

\S4method{dbConnect}{EXAConnection}(drv, ...)
}
\arguments{
\item{drv}{Existing connection}

\item{exahost}{Host and port of the database cluster, e.g. '10.0.2.15:8563'
or '10.0.2.15..20:8563'.}

\item{uid}{DB username, e.g. 'sys'}

\item{pwd}{DB user password, e.g. 'exasol'}

\item{schema}{Schema in EXASOL db which is opened directly after the
connection.}

\item{encryption}{Whether to use TLS encryption. 'Y' (default) enables TLS.}

\item{sslcertificate}{Controls server certificate verification when
\code{encryption == 'Y'}. Use \code{"SSL_VERIFY_NONE"} to disable server
verification, \code{"SSL_VERIFY_SERVER"} (or the default empty string) to
verify against the system trust store, or pass the path to a PEM
certificate file to verify against a specific trust anchor.}

\item{autocommit}{By default 'Y'. If 'Y' each SQL statement is committed. 'N'
means that no commits are executed automatically. The transaction will be
rolled back on disconnect, which causes the loss of all data written during
the transaction.}

\item{querytimeout}{Time in seconds the Exasol DB computes a query before it
is aborted. The default \code{'0'} (zero) means no timeout, i.e. queries
run until finished. Applied via the WebSocket \code{setAttributes} command.}

\item{...}{Additional parameters (currently unused).}
}
\value{
A fresh EXAConnection object.
}
\description{
dbConnect
}
\section{Methods (by class)}{
\itemize{
\item \code{dbConnect(EXADriver)}: Creates a new connection to an EXASOL Database via WebSocket.

\item \code{dbConnect(character)}: Creates a connection to an EXASOL Database based on string.

\item \code{dbConnect(EXAConnection)}: Creates a connection to an EXASOL Database based on an existing connection.

}}
\examples{
\dontrun{
 con <- dbConnect("exa", exahost = "212.209.123.20:8563",
                  uid = "peter", pwd = "password123", schema = "sales")
}
}
\seealso{
\code{\link[DBI:dbConnect]{DBI::dbConnect()}}

Other EXADriver related objects:
\code{\link{EXADriver-class}},
\code{\link[=dbEnd]{dbEnd()}},
\code{\link[=exasol_driver]{exasol_driver()}}

Other EXAConnection related objects:
\code{\link{EXAConnection-class}},
\code{\link{dbBegin,EXAConnection-method}},
\code{\link{dbCommit,EXAConnection-method}},
\code{\link[=dbCurrentSchema]{dbCurrentSchema()}},
\code{\link[=dbEnd]{dbEnd()}},
\code{\link{dbExistsTable,EXAConnection,character-method}},
\code{\link{dbGetQuery,EXAConnection,character-method}},
\code{\link{dbListFields,EXAConnection,character-method}},
\code{\link{dbListTables,EXAConnection-method}},
\code{\link{dbReadTable,EXAConnection,character-method}},
\code{\link{dbRemoveTable,EXAConnection,character-method}},
\code{\link{dbRollback,EXAConnection-method}},
\code{\link{dbSendQuery,EXAConnection,character-method}},
\code{\link{dbWriteTable,EXAConnection,character,data.frame-method}}
}
\author{
EXASOL AG <opensource@exasol.com>
}
\concept{EXAConnection related objects}
\concept{EXADriver related objects}
