% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXADBI-connection.R
\name{dbCurrentSchema}
\alias{dbCurrentSchema}
\title{Fetches and outputs the current schema from an EXASOL DB. Also updates EXAConnection metadata.}
\usage{
dbCurrentSchema(con, setSchema = NULL)
}
\arguments{
\item{con}{A valid EXAConnection}

\item{setSchema}{If not NULL, this schema will be opened and attached to EXAConnection metadata.}
}
\value{
an updated EXAConnection
}
\description{
Fetches and outputs the current schema from an EXASOL DB. Also updates EXAConnection metadata.
}
\seealso{
Other EXAConnection related objects:
\code{\link{EXAConnection-class}},
\code{\link{dbBegin,EXAConnection-method}},
\code{\link{dbCommit,EXAConnection-method}},
\code{\link{dbConnect,EXADriver-method}},
\code{\link[=dbEnd]{dbEnd()}},
\code{\link{dbExistsTable,EXAConnection,character-method}},
\code{\link{dbGetQuery,EXAConnection,character-method}},
\code{\link{dbListFields,EXAConnection,character-method}},
\code{\link{dbListTables,EXAConnection-method}},
\code{\link{dbReadTable,EXAConnection,character-method}},
\code{\link{dbRemoveTable,EXAConnection,character-method}},
\code{\link{dbRollback,EXAConnection-method}},
\code{\link{dbSendQuery,EXAConnection,character-method}},
\code{\link{dbWriteTable,EXAConnection,character,data.frame-method}}

Other None-DBI-API:
\code{\link[=dbEnd]{dbEnd()}},
\code{\link[=exa.createScript]{exa.createScript()}},
\code{\link[=exa.readData]{exa.readData()}},
\code{\link[=exa.writeData]{exa.writeData()}}
}
\concept{EXAConnection related objects}
\concept{None-DBI-API}
