% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXADBI-table.R
\name{dbExistsTable,EXAConnection,character-method}
\alias{dbExistsTable,EXAConnection,character-method}
\title{dbExistsTable}
\usage{
\S4method{dbExistsTable}{EXAConnection,character}(conn, name, schema = "")
}
\arguments{
\item{conn}{An EXAConnection object.}

\item{name}{A fully qualified table name in the form schema.table.}

\item{schema}{Alternatively to `name`, a schema can be specified separately.}
}
\value{
A logical indicating if the table exists.
}
\description{
dbExistsTable
}
\section{Methods (by class)}{
\itemize{
\item \code{dbExistsTable(conn = EXAConnection, name = character)}: Checks if a table exists in an EXASOL DB.

}}
\seealso{
\code{\link[DBI:dbExistsTable]{DBI::dbExistsTable()}}

Other EXAConnection related objects:
\code{\link{EXAConnection-class}},
\code{\link{dbBegin,EXAConnection-method}},
\code{\link{dbCommit,EXAConnection-method}},
\code{\link{dbConnect,EXADriver-method}},
\code{\link[=dbCurrentSchema]{dbCurrentSchema()}},
\code{\link[=dbEnd]{dbEnd()}},
\code{\link{dbGetQuery,EXAConnection,character-method}},
\code{\link{dbListFields,EXAConnection,character-method}},
\code{\link{dbListTables,EXAConnection-method}},
\code{\link{dbReadTable,EXAConnection,character-method}},
\code{\link{dbRemoveTable,EXAConnection,character-method}},
\code{\link{dbRollback,EXAConnection-method}},
\code{\link{dbSendQuery,EXAConnection,character-method}},
\code{\link{dbWriteTable,EXAConnection,character,data.frame-method}}

Other DQL functions:
\code{\link{dbClearResult,EXAResult-method}},
\code{\link{dbGetQuery,EXAConnection,character-method}},
\code{\link{dbListFields,EXAConnection,character-method}},
\code{\link{dbListTables,EXAConnection-method}},
\code{\link{dbReadTable,EXAConnection,character-method}},
\code{\link{dbSendQuery,EXAConnection,character-method}},
\code{\link{fetch,EXAResult,numeric-method}}
}
\concept{DQL functions}
\concept{EXAConnection related objects}
