% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXADBI.R
\docType{methods}
\name{dbFetch}
\alias{dbFetch}
\alias{fetch}
\title{Fetches a subset of an result set.}
\usage{
\S4method{fetch}{EXAResult,numeric}(res, n = -1, ...)
}
\arguments{
\item{res}{An EXAResult object.}

\item{n}{An int declaring the size of the subset to fetch. If missing, the whole subset is fetched.}

\item{...}{further arguments to be passed on to exa.readData.}
}
\description{
Fetches a subset of an result set.
}
\seealso{
Other DQL functions: \code{\link{dbClearResult}},
  \code{\link{dbExistsTable}}, \code{\link{dbGetQuery}},
  \code{\link{dbReadTable}}, \code{\link{dbSendQuery}}

Other EXAResult related objects: \code{\link{EXAResult-class}},
  \code{\link{dbClearResult}}
}

