% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXADBI-result.R
\name{dbHasCompleted,EXAResult-method}
\alias{dbHasCompleted,EXAResult-method}
\title{dbHasCompleted}
\usage{
\S4method{dbHasCompleted}{EXAResult}(res)
}
\arguments{
\item{res}{EXAResult object to query on.}
}
\description{
dbHasCompleted
}
\section{Methods (by class)}{
\itemize{
\item \code{dbHasCompleted(EXAResult)}: Returns if the operation has completed.

}}
\seealso{
\code{\link[DBI:dbHasCompleted]{DBI::dbHasCompleted()}}
}
\author{
EXASOL AG <opensource@exasol.com>
}
