% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXADBI-table.R
\name{dbListFields,EXAConnection,character-method}
\alias{dbListFields,EXAConnection,character-method}
\alias{dbListFields,EXAResult,ANY-method}
\title{dbListFields}
\usage{
\S4method{dbListFields}{EXAConnection,character}(conn, name, schema, ...)

\S4method{dbListFields}{EXAResult,ANY}(conn, name, ...)
}
\arguments{
\item{conn}{An EXAResult object}

\item{name}{A fully qualified table name in the form schema.table.}

\item{schema}{Filter on DB schema}

\item{...}{Not used.}
}
\value{
A character vector. If no tables present, a character vector of length 0.

A character vector. If no tables present, a character vector of length 0.
}
\description{
dbListFields
}
\section{Methods (by class)}{
\itemize{
\item \code{dbListFields(conn = EXAConnection, name = character)}: Lists all fields of a table.

\item \code{dbListFields(conn = EXAResult, name = ANY)}: Lists all fields of a result.

}}
\seealso{
\code{\link[DBI:dbListFields]{DBI::dbListFields()}}

Other EXAConnection related objects:
\code{\link{EXAConnection-class}},
\code{\link{dbBegin,EXAConnection-method}},
\code{\link{dbCommit,EXAConnection-method}},
\code{\link{dbConnect,EXADriver-method}},
\code{\link[=dbCurrentSchema]{dbCurrentSchema()}},
\code{\link[=dbEnd]{dbEnd()}},
\code{\link{dbExistsTable,EXAConnection,character-method}},
\code{\link{dbGetQuery,EXAConnection,character-method}},
\code{\link{dbListTables,EXAConnection-method}},
\code{\link{dbReadTable,EXAConnection,character-method}},
\code{\link{dbRemoveTable,EXAConnection,character-method}},
\code{\link{dbRollback,EXAConnection-method}},
\code{\link{dbSendQuery,EXAConnection,character-method}},
\code{\link{dbWriteTable,EXAConnection,character,data.frame-method}}

Other DQL functions:
\code{\link{dbClearResult,EXAResult-method}},
\code{\link{dbExistsTable,EXAConnection,character-method}},
\code{\link{dbGetQuery,EXAConnection,character-method}},
\code{\link{dbListTables,EXAConnection-method}},
\code{\link{dbReadTable,EXAConnection,character-method}},
\code{\link{dbSendQuery,EXAConnection,character-method}},
\code{\link{fetch,EXAResult,numeric-method}}

Other EXAResult related objects:
\code{\link{EXAResult-class}},
\code{\link{dbClearResult,EXAResult-method}},
\code{\link{fetch,EXAResult,numeric-method}}
}
\concept{DQL functions}
\concept{EXAConnection related objects}
\concept{EXAResult related objects}
