% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXADBI-table.R
\name{dbReadTable,EXAConnection,character-method}
\alias{dbReadTable,EXAConnection,character-method}
\title{dbReadTable}
\usage{
\S4method{dbReadTable}{EXAConnection,character}(conn, name, schema = "", order_col = NA, limit = NA, ...)
}
\arguments{
\item{conn}{An EXAConnection object.}

\item{name}{A fully qualified table name in the form schema.table.}

\item{schema}{Alternatively and with preference to `name`, a schema can be specified separately.}

\item{order_col}{A string containing columns to have the result set ordered upon, e.g. "col1 desc, col2"}

\item{limit}{A row limit to the result set.
CAUTION: a limit without order clause is non-deterministic in EXASOL,
i.e. subsequent runs of the same statement may deliver differing result sets.}

\item{...}{Further arguments to be passed on to exa.readData.}
}
\value{
The result exa.readData, by default a data.frame containing the result set.
}
\description{
dbReadTable
}
\section{Functions}{
\itemize{
\item \code{dbReadTable(conn = EXAConnection, name = character)}: Reads a DB table.

}}
\seealso{
\code{\link[DBI:dbReadTable]{DBI::dbReadTable()}}

Other EXAConnection related objects:
\code{\link{EXAConnection-class}},
\code{\link{dbBegin,EXAConnection-method}},
\code{\link{dbCommit,EXAConnection-method}},
\code{\link{dbConnect,EXADriver-method}},
\code{\link[=dbCurrentSchema]{dbCurrentSchema()}},
\code{\link[=dbEnd]{dbEnd()}},
\code{\link{dbExistsTable,EXAConnection,character-method}},
\code{\link{dbGetQuery,EXAConnection,character-method}},
\code{\link{dbListFields,EXAConnection,character-method}},
\code{\link{dbListTables,EXAConnection-method}},
\code{\link{dbRemoveTable,EXAConnection,character-method}},
\code{\link{dbRollback,EXAConnection-method}},
\code{\link{dbSendQuery,EXAConnection,character-method}},
\code{\link{dbWriteTable,EXAConnection,character,data.frame-method}}

Other DQL functions:
\code{\link{dbClearResult,EXAResult-method}},
\code{\link{dbExistsTable,EXAConnection,character-method}},
\code{\link{dbGetQuery,EXAConnection,character-method}},
\code{\link{dbListFields,EXAConnection,character-method}},
\code{\link{dbListTables,EXAConnection-method}},
\code{\link{dbSendQuery,EXAConnection,character-method}},
\code{\link{fetch,EXAResult,numeric-method}}
}
\concept{DQL functions}
\concept{EXAConnection related objects}
