% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXADBI-table.R
\name{dbRemoveTable,EXAConnection,character-method}
\alias{dbRemoveTable,EXAConnection,character-method}
\title{dbRemoveTable}
\usage{
\S4method{dbRemoveTable}{EXAConnection,character}(conn, name, schema = "", cascade = FALSE)
}
\arguments{
\item{conn}{An EXAConnection object.}

\item{name}{A fully qualified table identifier (schema.table) or a simple table identifier, if the
parameter schema is specified.}

\item{schema}{An optional string specifying the DB schema. Caution: use either this parameter, or give
a fully qualified identifier as giving both will result in an error.}

\item{cascade}{A logical indicating whether also foreign key constraints referencing the table should
be deleted. Default is FALSE.}
}
\value{
A logical indicating success.
}
\description{
dbRemoveTable
}
\section{Methods (by class)}{
\itemize{
\item \code{dbRemoveTable(conn = EXAConnection, name = character)}: Removes a table.

}}
\seealso{
\code{\link[DBI:dbRemoveTable]{DBI::dbRemoveTable()}}

Other EXAConnection related objects:
\code{\link{EXAConnection-class}},
\code{\link{dbBegin,EXAConnection-method}},
\code{\link{dbCommit,EXAConnection-method}},
\code{\link{dbConnect,EXADriver-method}},
\code{\link[=dbCurrentSchema]{dbCurrentSchema()}},
\code{\link[=dbEnd]{dbEnd()}},
\code{\link{dbExistsTable,EXAConnection,character-method}},
\code{\link{dbGetQuery,EXAConnection,character-method}},
\code{\link{dbListFields,EXAConnection,character-method}},
\code{\link{dbListTables,EXAConnection-method}},
\code{\link{dbReadTable,EXAConnection,character-method}},
\code{\link{dbRollback,EXAConnection-method}},
\code{\link{dbSendQuery,EXAConnection,character-method}},
\code{\link{dbWriteTable,EXAConnection,character,data.frame-method}}

Other DML functions:
\code{\link{dbWriteTable,EXAConnection,character,data.frame-method}}
}
\author{
EXASOL AG <opensource@exasol.com>
}
\concept{DML functions}
\concept{EXAConnection related objects}
