% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXADBI-transaction.R
\name{dbRollback,EXAConnection-method}
\alias{dbRollback,EXAConnection-method}
\title{dbRollback}
\usage{
\S4method{dbRollback}{EXAConnection}(conn)
}
\arguments{
\item{conn}{An EXAConnection object}
}
\value{
A logical indicating success.
}
\description{
dbRollback
}
\section{Methods (by class)}{
\itemize{
\item \code{dbRollback(EXAConnection)}: Rolls the current DB transaction back.

}}
\seealso{
\code{\link[DBI:dbRollback]{DBI::dbRollback()}}

Other EXAConnection related objects:
\code{\link{EXAConnection-class}},
\code{\link{dbBegin,EXAConnection-method}},
\code{\link{dbCommit,EXAConnection-method}},
\code{\link{dbConnect,EXADriver-method}},
\code{\link[=dbCurrentSchema]{dbCurrentSchema()}},
\code{\link[=dbEnd]{dbEnd()}},
\code{\link{dbExistsTable,EXAConnection,character-method}},
\code{\link{dbGetQuery,EXAConnection,character-method}},
\code{\link{dbListFields,EXAConnection,character-method}},
\code{\link{dbListTables,EXAConnection-method}},
\code{\link{dbReadTable,EXAConnection,character-method}},
\code{\link{dbRemoveTable,EXAConnection,character-method}},
\code{\link{dbSendQuery,EXAConnection,character-method}},
\code{\link{dbWriteTable,EXAConnection,character,data.frame-method}}

Other transaction management functions:
\code{\link{dbBegin,EXAConnection-method}},
\code{\link{dbCommit,EXAConnection-method}},
\code{\link[=dbEnd]{dbEnd()}}
}
\author{
EXASOL AG <opensource@exasol.com>
}
\concept{EXAConnection related objects}
\concept{transaction management functions}
