% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXADBI-table.R
\name{dbWriteTable,EXAConnection,character,data.frame-method}
\alias{dbWriteTable,EXAConnection,character,data.frame-method}
\title{dbWriteTable}
\usage{
\S4method{dbWriteTable}{EXAConnection,character,data.frame}(
  conn,
  name,
  value,
  schema = "",
  field_types,
  overwrite = FALSE,
  writeCols = NA,
  ...
)
}
\arguments{
\item{conn}{An EXAConnection object.}

\item{name}{A fully qualified table name (schema.table). Alternatively the schema can be given
via the parameter 'schema' (see below).}

\item{value}{A data.frame containing data.}

\item{schema}{A schema identifier.}

\item{field_types}{A character vector containing the column data types, in the form
of c("varchar(20)","int").
   If missing, the column types of the data.frame are being converted and used.}

\item{overwrite}{A logical indicating if existing data shall be overwritten. Default is 'FALSE',
i.e. new data is appended to the DB table.}

\item{writeCols}{A logical or a character vector containing the cols of the DB table to be
written into, in the form of c("col1","col4","col3"). If set to TRUE, then the column names
of the data.frame are used.
If FALSE, NA, or missing, no write columns are defined and columns are matched by column order.
Default is NA. Useful to change if the DB table contains more columns than the data.frame, or
if the column order differs.}

\item{...}{Additional parameters to be passed on to exa.writeData.}
}
\value{
A logical indicating success.
}
\description{
dbWriteTable
}
\section{Functions}{
\itemize{
\item \code{dbWriteTable(conn = EXAConnection, name = character, value = data.frame)}: Writes a data.frame into a table. If the table does not exist, it is created.

}}
\seealso{
\code{\link[DBI:dbWriteTable]{DBI::dbWriteTable()}}

Other EXAConnection related objects:
\code{\link{EXAConnection-class}},
\code{\link{dbBegin,EXAConnection-method}},
\code{\link{dbCommit,EXAConnection-method}},
\code{\link{dbConnect,EXADriver-method}},
\code{\link[=dbCurrentSchema]{dbCurrentSchema()}},
\code{\link[=dbEnd]{dbEnd()}},
\code{\link{dbExistsTable,EXAConnection,character-method}},
\code{\link{dbGetQuery,EXAConnection,character-method}},
\code{\link{dbListFields,EXAConnection,character-method}},
\code{\link{dbListTables,EXAConnection-method}},
\code{\link{dbReadTable,EXAConnection,character-method}},
\code{\link{dbRemoveTable,EXAConnection,character-method}},
\code{\link{dbRollback,EXAConnection-method}},
\code{\link{dbSendQuery,EXAConnection,character-method}}

Other DML functions:
\code{\link{dbRemoveTable,EXAConnection,character-method}}
}
\concept{DML functions}
\concept{EXAConnection related objects}
