% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXADBI-connection.R
\docType{class}
\name{EXAConnection-class}
\alias{EXAConnection-class}
\alias{EXAConnection}
\title{An Object holding a connection to an EXASOL Database.}
\description{
An Object holding a connection to an EXASOL Database.
}
\section{Slots}{

\describe{
\item{\code{ws_handle}}{An external pointer to the C++ WebSocket session.}

\item{\code{session_id}}{An integer containing the Exasol session ID.}

\item{\code{current_schema}}{A string reflecting the current schema.}

\item{\code{autocom_default}}{A logical indicating if autocommit is active.}

\item{\code{db_host}}{A string containing the hostname or IP.}

\item{\code{db_port}}{An integer containing the connection port.}

\item{\code{db_user}}{A string containing the database user name.}

\item{\code{db_name}}{A string containing the database name.}

\item{\code{db_prod_name}}{A string containing the database product name.}

\item{\code{db_version}}{A string containing the database version.}

\item{\code{drv_name}}{A string containing the connection driver version.}

\item{\code{encrypted}}{A logical indicating if connection should be encrypted.}

\item{\code{sslcertificate}}{The TLS verification setting used at connect time
(\code{""}, \code{"SSL_VERIFY_NONE"}, \code{"SSL_VERIFY_SERVER"} or a path
to a PEM trust file). Used to reproduce the original TLS behavior when the
connection is cloned.}
}}

\seealso{
\code{\link{DBIConnection-class}}

Other DBI classes:
\code{\link{EXADriver-class}},
\code{\link{EXAObject-class}},
\code{\link{EXAResult-class}}

Other EXAConnection related objects:
\code{\link{dbBegin,EXAConnection-method}},
\code{\link{dbCommit,EXAConnection-method}},
\code{\link{dbConnect,EXADriver-method}},
\code{\link[=dbCurrentSchema]{dbCurrentSchema()}},
\code{\link[=dbEnd]{dbEnd()}},
\code{\link{dbExistsTable,EXAConnection,character-method}},
\code{\link{dbGetQuery,EXAConnection,character-method}},
\code{\link{dbListFields,EXAConnection,character-method}},
\code{\link{dbListTables,EXAConnection-method}},
\code{\link{dbReadTable,EXAConnection,character-method}},
\code{\link{dbRemoveTable,EXAConnection,character-method}},
\code{\link{dbRollback,EXAConnection-method}},
\code{\link{dbSendQuery,EXAConnection,character-method}},
\code{\link{dbWriteTable,EXAConnection,character,data.frame-method}}
}
\author{
EXASOL AG <opensource@exasol.com>
}
\concept{DBI classes}
\concept{EXAConnection related objects}
