% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exa.log.R
\name{EXAConnectionLog}
\alias{EXAConnectionLog}
\title{EXAConnectionLog
Shows the EXASOL connection log.

This function is no longer supported after the migration from ODBC to WebSocket.
It previously relied on the ODBC connection string to find the log file.
Use EXATrace to enable C-level tracing instead.}
\usage{
EXAConnectionLog(con, view = TRUE)
}
\arguments{
\item{con}{An EXAConnection object}

\item{view}{logical If true, the log is shown in R View, otherwise, a data.frame is returned.}
}
\value{
a data.frame, if View is FALSE.
}
\description{
EXAConnectionLog
Shows the EXASOL connection log.

This function is no longer supported after the migration from ODBC to WebSocket.
It previously relied on the ODBC connection string to find the log file.
Use EXATrace to enable C-level tracing instead.
}
