% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exasol.R
\docType{package}
\name{exasol-package}
\alias{exasol-package}
\alias{exasol}
\title{R Interface & SDK for the EXASOL Database}
\description{
The EXASOL R Package offers interface functionality such as connecting to, querying
and writing into an EXASOL Database (version 7.1 onwards) via a native WebSocket protocol.
No ODBC driver installation is required. It is optimised for massively parallel reading & writing
from and to a multinode cluster using HTTP/HTTPS bulk transfer channels. Implemented are DBI
compliant methods for database access, querying and modification. The package integrates with
EXASOL's InDB R UDF capabilities, which allows to deploy and execute R code dynamically from an
R application running on a client.

EXASOL is an InMemory RDBMS that runs in a MPP cluster (shared-nothing) environment. Leading the TPC-H
benchmark, it is considered the fastest analytical data warehouse available. The community edition
can be downloaded for free on \url{https://docs.exasol.com/get_started/communityedition.htm}.
}
\section{Package index}{


\subsection{DBI: General}{

\describe{
\item{\code{\link{EXAObject-class}}}{A virtual meta Object for all other EXASOL DBI objects.}
\item{\code{\link{dbGetInfo}}}{}
\item{\code{\link{summary}}}{}
\item{\code{\link{dbIsValid}}}{Checks if an EXAObject is still valid.}
\item{\code{\link{dbDataType}}}{Determine the EXASOL data type of an object.}
}
}

\subsection{DBI: Driver-related}{

\describe{
\item{\code{\link{EXADriver-class}}}{An interface driver object to the EXASOL Database.}
\item{\code{\link{dbGetInfo}}}{}
\item{\code{\link{summary}}}{}
\item{\code{\link{dbDriver}}}{Load database drivers.}
\item{\code{\link{dbUnloadDriver}}}{Unload a driver.}
\item{\code{\link{dbConnect}}}{Creates a connection to an EXASOL Database.}
}
}

\subsection{DBI: Connection-related}{
\describe{
\item{\code{\link{EXAConnection-class}}}{An object holding a connection to an EXASOL Database.}
\item{\code{\link{dbGetInfo}}}{}
\item{\code{\link{summary}}}{}
\item{\code{\link{dbDisconnect}}}{Disconnects the connection.}
\item{\code{\link{dbSendQuery}}}{Sends an SQL statment to an EXASOL DB, prepares for result fetching.}
\item{\code{\link{dbGetQuery}}}{Executes the query, fetches and returns the entire result set.}
\item{\code{\link{dbGetException}}}{}
\item{\code{\link{dbListResults}}}{}
\item{\code{\link{dbListFields}}}{}
\item{\code{\link{dbListTables}}}{}
\item{\code{\link{dbReadTable}}}{Reads a DB table.}
\item{\code{\link{dbWriteTable}}}{Writes a data.frame into a table. If the table does not exist, it is created.}
\item{\code{\link{dbExistsTable}}}{Checks if a table exists in an EXASOL DB.}
\item{\code{\link{dbRemoveTable}}}{Removes a table.}
\item{\code{\link{dbBegin}}}{Starts a DB transaction.}
\item{\code{\link{dbEnd}}}{Ends a DB transaction.}
\item{\code{\link{dbCommit}}}{Sends a commit.}
\item{\code{\link{dbRollback}}}{Rolls the current DB transaction back.}
}}

\subsection{DBI: Result set-related}{
\describe{
\item{\code{\link{EXAResult-class}}}{An object that is associated with a result set in an EXASOL Database.}
\item{\code{\link{dbGetInfo}}}{}
\item{\code{\link{summary}}}{}
\item{\code{\link{dbFetch}}}{Fetches a subset of an result set.}
\item{\code{\link{dbClearResult}}}{Frees all resources associated with an \code{EXAResult}.}
\item{\code{\link{dbColumnInfo}}}{}
\item{\code{\link{dbGetStatement}}}{}
\item{\code{\link{dbHasCompleted}}}{}
\item{\code{\link{dbGetRowsAffected}}}{}
\item{\code{\link{dbGetRowCount}}}{}
}}

\subsection{Low-level methods}{
\describe{
\item{\code{\link{exa.readData}}}{Execute a SQL query on an EXASolution database and read results fast.}
\item{\code{\link{exa.writeData}}}{Write a data.frame into an EXASOL table fast.}
\item{\code{\link{exa.createScript}}}{Deploys an R function as an UDF in the EXASolution database.}
\item{\code{\link{dbCurrentSchema}}}{Fetches and outputs the current schema from an EXASOL DB.}
}}
}

\references{
\enumerate{
\item The development version of the package is available on \url{https://github.com/EXASOL/r-exasol}
\item Bugs and improvements may be noted on \url{https://github.com/EXASOL/r-exasol/issues}
\item Downloads & manuals related to the EXASOL Database are at \url{https://docs.exasol.com/get_started/communityedition.htm}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/EXASOL/r-exasol} \url{https://www.exasol.com}
  \item Report bugs at \url{https://github.com/exasol/r-exasol/issues}
}

}
\author{
EXASOL AG & Community

Maintainer: \packageMaintainer{exasol}
}
\keyword{distributed}
\keyword{in-memory}
\keyword{sql}
