% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXADBI-driver.R
\name{exasol_driver}
\alias{exasol_driver}
\alias{exa}
\title{Instantiates an EXADriver object.}
\usage{
exasol_driver(silent = FALSE)

exa(silent = FALSE)
}
\arguments{
\item{silent}{If TRUE, no message is printed.}
}
\value{
An EXADriver object.
}
\description{
Instantiates an EXADriver object.
}
\section{Functions}{
\itemize{
\item \code{exa()}: An alias for \code{\link{exasol_driver}}

}}
\seealso{
Other EXADriver related objects:
\code{\link{EXADriver-class}},
\code{\link{dbConnect,EXADriver-method}},
\code{\link[=dbEnd]{dbEnd()}}
}
\author{
EXASOL AG <opensource@exasol.com>
}
\concept{EXADriver related objects}
