% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXADBI.R
\name{EXATrace}
\alias{EXATrace}
\title{Enables tracing for debugging purpose.}
\usage{
EXATrace(tracefile)
}
\arguments{
\item{tracefile}{The output tracefile}
}
\description{
Enables tracing for debugging purpose.
}
