% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXADBI.R
\name{EXAupper}
\alias{EXAupper}
\title{Changes an identifier into uppercase, except for it is quoted.}
\usage{
EXAupper(identifier)
}
\arguments{
\item{identifier}{A character vector containing one or many (table, schema, column,...) identifiers.}
}
\value{
A character vector containing one or many processed identifiers.
}
\description{
Changes an identifier into uppercase, except for it is quoted.
}

