% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXADBI-internal.R
\name{grep_schema}
\alias{grep_schema}
\title{Tries to extrapolate the schema name from the statement given. It looks for strings framed by 'from' and a dot (not case sensitive).}
\usage{
grep_schema(stmt, statement = TRUE)
}
\arguments{
\item{stmt}{The statement used}
}
\value{
A character vector containing all schemas found, or alt
}
\description{
Tries to extrapolate the schema name from the statement given. It looks for strings framed by 'from' and a dot (not case sensitive).
}
\keyword{internal}
