% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXADBI-internal.R
\name{processIDs}
\alias{processIDs}
\title{Takes an identifier, and transforms it into an EXASOL compatible identifier. In detail: it
 - removes surrounding whitespaces
 - doubles quotes inside the string,
 - changes dots to underscores,
 - adds surrounding quotes.
 param id A character vector containing identifiers.
 param quotes The kind of quotes to be used: double-qoutes for identifiers with special chars, single
 quotes for strings.
 return A list of character vectors, each containing schema and table identifier. conversion
 via as.data.frame fails for the method strips escaped quotes.}
\usage{
processIDs(id, quotes = "\\"")
}
\description{
Takes an identifier, and transforms it into an EXASOL compatible identifier. In detail: it
 - removes surrounding whitespaces
 - doubles quotes inside the string,
 - changes dots to underscores,
 - adds surrounding quotes.
 param id A character vector containing identifiers.
 param quotes The kind of quotes to be used: double-qoutes for identifiers with special chars, single
 quotes for strings.
 return A list of character vectors, each containing schema and table identifier. conversion
 via as.data.frame fails for the method strips escaped quotes.
}
\keyword{internal}
