% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exasol.R
\name{RETURNS}
\alias{RETURNS}
\title{RETURNS output type of UDF script -- function emits just a single value.}
\usage{
RETURNS
}
\description{
RETURNS output type of UDF script -- function emits just a single value.
}
