% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXADBI-connection.R
\name{RODBC-class}
\alias{RODBC-class}
\title{The S3 class RODBC will be registered as a superclass of EXAConnection}
\description{
The S3 class RODBC will be registered as a superclass of EXAConnection
}
