% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exasol.R
\name{SCALAR}
\alias{SCALAR}
\title{SCALAR input type of UDF script will call the function once for each record.}
\usage{
SCALAR
}
\description{
SCALAR input type of UDF script will call the function once for each record.
}
