% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exasol.R
\docType{data}
\name{SET}
\alias{SET}
\title{SET input type of UDF script will call the function once for each group}
\format{\preformatted{ symbol SET
}}
\usage{
SET
}
\description{
SET input type of UDF script will call the function once for each group
}
\keyword{datasets}

