% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXADBI-summary.R
\name{summary,EXAObject-method}
\alias{summary,EXAObject-method}
\alias{summary,EXAResult-method}
\alias{summary,EXADriver-method}
\alias{summary,EXAConnection-method}
\title{summary}
\usage{
\S4method{summary}{EXAObject}(object, ...)

\S4method{summary}{EXAResult}(object, ...)

\S4method{summary}{EXADriver}(object, ...)

\S4method{summary}{EXAConnection}(object, ...)
}
\arguments{
\item{object}{EXAConnection object for which the summary info will be returned.}

\item{...}{Further parameters passed to summary().}
}
\value{
A summary string of the EXAObject

A summary string of the EXAResult

A summary string of the EXADriver

A summary string of the EXADriver
}
\description{
summary
}
\section{Methods (by class)}{
\itemize{
\item \code{summary(EXAObject)}: Summary info for EXASOL object

\item \code{summary(EXAResult)}: Summary info for EXASOL EXAResult

\item \code{summary(EXADriver)}: Summary info for EXASOL EXADriver

\item \code{summary(EXAConnection)}: Info for EXASOL EXAConnection.

}}
\author{
EXASOL AG <opensource@exasol.com>
}
