% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exasol.R
\docType{data}
\name{C_asyncRODBCIOStart}
\alias{C_asyncRODBCIOStart}
\title{Prepares the highspeed data channel. Check the developer_guide for more information.}
\format{
An object of class \code{CallRoutine} (inherits from \code{NativeSymbolInfo}) of length 4.
}
\usage{
C_asyncRODBCIOStart
}
\arguments{
\item{hostA}{Database host address}

\item{portA}{Database port}

\item{protocolA}{Http or Https}
}
\description{
Prepares the highspeed data channel. Check the developer_guide for more information.
}
\keyword{internal}
