% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXADBI.R
\docType{methods}
\name{dbBegin}
\alias{dbBegin}
\title{Starts a DB transaction. In EXASOL, it disables autocommit.}
\usage{
\S4method{dbBegin}{EXAConnection}(conn)
}
\arguments{
\item{conn}{An EXAConnection object}
}
\value{
a logical indicating success.
}
\description{
Starts a DB transaction. In EXASOL, it disables autocommit.
}
\seealso{
Other EXAConnection related objects: \code{\link{EXAConnection-class}},
  \code{\link{dbCommit}}, \code{\link{dbConnect}},
  \code{\link{dbCurrentSchema}}, \code{\link{dbEnd}},
  \code{\link{dbExistsTable}}, \code{\link{dbGetQuery}},
  \code{\link{dbReadTable}}, \code{\link{dbRemoveTable}},
  \code{\link{dbRollback}}, \code{\link{dbSendQuery}},
  \code{\link{dbWriteTable}}

Other transaction management functions: \code{\link{dbCommit}},
  \code{\link{dbEnd}}, \code{\link{dbRollback}}
}

