% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXADBI.R
\name{dbClearResult}
\alias{dbClearResult}
\title{Frees all resources associated with an \code{EXAResult}.}
\usage{
\S4method{dbClearResult}{EXAResult}(res, ...)
}
\arguments{
\item{res}{An EXAResult object.}

\item{...}{Further arguments to passed to res$close(). This may be 'commit=TRUE' (not advisable).}
}
\value{
A logical indicating success.
}
\description{
Frees all resources associated with an \code{EXAResult}.
}
\seealso{
Other EXAResult related objects: 
\code{\link{EXAResult-class}},
\code{\link{dbFetch}()}

Other DQL functions: 
\code{\link{dbExistsTable}()},
\code{\link{dbFetch}()},
\code{\link{dbGetQuery}()},
\code{\link{dbListFields}()},
\code{\link{dbListTables}()},
\code{\link{dbReadTable}()},
\code{\link{dbSendQuery}()}
}
\concept{DQL functions}
\concept{EXAResult related objects}
