% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXADBI.R
\docType{methods}
\name{dbCommit}
\alias{dbCommit}
\title{Sends a commit.}
\usage{
\S4method{dbCommit}{EXAConnection}(conn, silent = FALSE)
}
\arguments{
\item{conn}{An EXAConnection object}
}
\value{
a logical indicating success.
}
\description{
Sends a commit.
}
\seealso{
Other EXAConnection related objects: \code{\link{EXAConnection-class}},
  \code{\link{dbBegin}}, \code{\link{dbConnect}},
  \code{\link{dbCurrentSchema}}, \code{\link{dbEnd}},
  \code{\link{dbExistsTable}}, \code{\link{dbGetQuery}},
  \code{\link{dbListFields}}, \code{\link{dbListTables}},
  \code{\link{dbReadTable}}, \code{\link{dbRemoveTable}},
  \code{\link{dbRollback}}, \code{\link{dbSendQuery}},
  \code{\link{dbWriteTable}}

Other transaction management functions: \code{\link{dbBegin}},
  \code{\link{dbEnd}}, \code{\link{dbRollback}}
}
