% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXADBI.R
\docType{methods}
\name{dbConnect}
\alias{dbConnect}
\title{Creates a connection to an EXASOL Database.}
\usage{
\S4method{dbConnect}{EXADriver}(drv, ...)
}
\arguments{
\item{drv}{An EXAdriver object, a character string "exasol" or "exa", or an existing
EXAConnection object (for connection cloning).}

\item{...}{Additional parameters to the connection string. If a connection is cloned, these override the old connection settings.}

\item{exahost}{DNS or IP and port of the database cluster, e.g. '10.0.2.15..20:8563'}

\item{uid}{DB username, e.g. 'sys'}

\item{pwd}{DB user password, e.g. 'exasol'}

\item{schema}{Schema in EXASOL db which is opened directly after the connection.}

\item{exalogfile}{the EXASOL odbc driver log file. By standard a tempfile is created. Log data may be
accessed with 'EXAlog(EXAConnection)'.}

\item{logmode}{EXASOL ODBC driver log mode. By default, no log is written ('NONE'). Other options are
'DEFAULT'           (most important function calls & SQL commands),
'VERBOSE'           (also additional data about internal steps & result data),
'ON ERROR ONLY'     (only errors are logged),
'DEBUGCOMM'         (extended logs, similar to verbose but w/o data & parameter tables).}

\item{encryption}{ODBC encryption. By default off. Switch on with 'Y'.}

\item{autocommit}{By default 'N'. If Y' each SQL statement is committed. Off means that
no commits are executed automatically. The transaction will be rolled back on disconnect, which
causes the loss of all data written during the transaction. The methods in the EXASOL package
handle their transaction themselves, but but if the connection is used otherwise it might be advisable to
switch autocommit to 'Y'.}

\item{querytimeout}{Time EXASOL DB computes a query before it is aborted. The default '0' (zero) means
no timeout, i.e. runs until finished.}

\item{connectionlcctype}{Sets the connection locale LC CTYPE. The default is the setting of the client's current R session.}

\item{connectionlcnumeric}{Sets the connection locale LC NUMERIC. The default is the setting of the client's current R session.}

\item{encoding}{The connection encoding. TODO.}

\item{dsn}{A preconfigured ODBC Data Source Name. Parameter being evaluated with priority to EXAHOST.}

\item{connection_string}{alternatively to everything else, a custom ODBC connection sting can be
provided. See EXASOL DB manual secion 4.2.5 for details, available at \url{https://www.exasol.com/portal}.}
}
\value{
a fresh EXAConnection object.
}
\description{
Creates a connection to an EXASOL Database.
}
\seealso{
Other EXAConnection related objects: \code{\link{EXAConnection-class}},
  \code{\link{dbBegin}}, \code{\link{dbCommit}},
  \code{\link{dbCurrentSchema}}, \code{\link{dbEnd}},
  \code{\link{dbExistsTable}}, \code{\link{dbGetQuery}},
  \code{\link{dbReadTable}}, \code{\link{dbRemoveTable}},
  \code{\link{dbRollback}}, \code{\link{dbSendQuery}},
  \code{\link{dbWriteTable}}

Other EXADriver related objects: \code{\link{EXADriver-class}}
}

