% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXADBI-connection.R
\name{dbConnect,EXADriver-method}
\alias{dbConnect,EXADriver-method}
\alias{dbConnect,character-method}
\alias{dbConnect,EXAConnection-method}
\title{dbConnect}
\usage{
\S4method{dbConnect}{EXADriver}(
  drv,
  exahost = "",
  uid = "",
  pwd = "",
  schema = "SYS",
  exalogfile = tempfile(pattern = "EXAODBC_", fileext = ".log"),
  logmode = "NONE",
  encryption = "N",
  autocommit = "Y",
  querytimeout = "0",
  connectionlcctype = Sys.getlocale(category = "LC_CTYPE"),
  connectionlcnumeric = Sys.getlocale(category = "LC_NUMERIC"),
  ...,
  dsn = "",
  connection_string = ""
)

\S4method{dbConnect}{character}(drv, ...)

\S4method{dbConnect}{EXAConnection}(drv, ...)
}
\arguments{
\item{drv}{Existing connection}

\item{exahost}{DNS or IP (or range of IPs) and port of the database cluster,
e.g. '10.0.2.15..20:8563'}

\item{uid}{DB username, e.g. 'sys'}

\item{pwd}{DB user password, e.g. 'exasol'}

\item{schema}{Schema in EXASOL db which is opened directly after the
connection.}

\item{exalogfile}{the EXASOL odbc driver log file. By standard a tempfile is
created. Log data may be accessed with 'EXAlog(EXAConnection)'.}

\item{logmode}{EXASOL ODBC driver log mode. Allowed options are:
\describe{
 \item{NONE}{no log is written (default)}
 \item{DEFAULT}{most important function calls & SQL commands}
 \item{VERBOSE}{also additional data about internal steps & result data}
 \item{ON ERROR ONLY}{only errors are logged}
 \item{DEBUGCOMM}{extended logs, similar to verbose but w/o data & parameter
 tables}
}}

\item{encryption}{ODBC encryption. By default off. Switch on with 'Y'.}

\item{autocommit}{By default 'Y'. If Y' each SQL statement is committed. 'N'
means that no commits are executed automatically. The transaction will be
rolled back on disconnect, which causes the loss of all data written during
the transaction.}

\item{querytimeout}{Time EXASOL DB computes a query before it is aborted.
The default \code{'0'} (zero) means no timeout, i.e. runs until finished.}

\item{connectionlcctype}{Sets the connection locale \code{LC CTYPE}.
The default is the setting of the client's current R session.}

\item{connectionlcnumeric}{Sets the connection locale \code{LC NUMERIC}.
The default is the setting of the client's current R session.}

\item{...}{Additional parameters to the connection string. If a connection is
cloned, these override the old connection settings.}

\item{dsn}{A preconfigured ODBC Data Source Name. Parameter being evaluated
with priority to \code{EXAHOST}.}

\item{connection_string}{alternatively to everything else, a custom ODBC
connection sting can be provided. See EXASOL DB manual secion 4.2.5 for
details, available at \url{https://docs.exasol.com/}.}
}
\value{
A fresh EXAConnection object.
}
\description{
dbConnect
}
\section{Methods (by class)}{
\itemize{
\item \code{EXADriver}: Creates a new connection to an EXASOL Database.

\item \code{character}: Creates a connection to an EXASOL Database based on string.

\item \code{EXAConnection}: Creates a connection to an EXASOL Database based on an existing connection.
}}

\examples{
\dontrun{
 con <- dbConnect("exa", dsn = "EXASolo")
 con <- dbConnect("exa", exahost = "212.209.123.20..25:8563",
                  uid = "peter", pwd = "password123", schema = "sales")
}
}
\seealso{
\code{\link[DBI:dbConnect]{DBI::dbConnect()}}

Other EXADriver related objects: 
\code{\link{EXADriver-class}},
\code{\link{dbEnd}()},
\code{\link{exasol_driver}()}

Other EXAConnection related objects: 
\code{\link{EXAConnection-class}},
\code{\link{dbBegin,EXAConnection-method}},
\code{\link{dbCommit,EXAConnection-method}},
\code{\link{dbCurrentSchema}()},
\code{\link{dbEnd}()},
\code{\link{dbExistsTable,EXAConnection,character-method}},
\code{\link{dbGetQuery,EXAConnection,character-method}},
\code{\link{dbListFields,EXAConnection,character-method}},
\code{\link{dbListTables,EXAConnection-method}},
\code{\link{dbReadTable,EXAConnection,character-method}},
\code{\link{dbRemoveTable,EXAConnection,character-method}},
\code{\link{dbRollback,EXAConnection-method}},
\code{\link{dbSendQuery,EXAConnection,character-method}},
\code{\link{dbWriteTable,EXAConnection,character,data.frame-method}}

Other EXADriver related objects: 
\code{\link{EXADriver-class}},
\code{\link{dbEnd}()},
\code{\link{exasol_driver}()}

Other EXAConnection related objects: 
\code{\link{EXAConnection-class}},
\code{\link{dbBegin,EXAConnection-method}},
\code{\link{dbCommit,EXAConnection-method}},
\code{\link{dbCurrentSchema}()},
\code{\link{dbEnd}()},
\code{\link{dbExistsTable,EXAConnection,character-method}},
\code{\link{dbGetQuery,EXAConnection,character-method}},
\code{\link{dbListFields,EXAConnection,character-method}},
\code{\link{dbListTables,EXAConnection-method}},
\code{\link{dbReadTable,EXAConnection,character-method}},
\code{\link{dbRemoveTable,EXAConnection,character-method}},
\code{\link{dbRollback,EXAConnection-method}},
\code{\link{dbSendQuery,EXAConnection,character-method}},
\code{\link{dbWriteTable,EXAConnection,character,data.frame-method}}

Other EXADriver related objects: 
\code{\link{EXADriver-class}},
\code{\link{dbEnd}()},
\code{\link{exasol_driver}()}

Other EXAConnection related objects: 
\code{\link{EXAConnection-class}},
\code{\link{dbBegin,EXAConnection-method}},
\code{\link{dbCommit,EXAConnection-method}},
\code{\link{dbCurrentSchema}()},
\code{\link{dbEnd}()},
\code{\link{dbExistsTable,EXAConnection,character-method}},
\code{\link{dbGetQuery,EXAConnection,character-method}},
\code{\link{dbListFields,EXAConnection,character-method}},
\code{\link{dbListTables,EXAConnection-method}},
\code{\link{dbReadTable,EXAConnection,character-method}},
\code{\link{dbRemoveTable,EXAConnection,character-method}},
\code{\link{dbRollback,EXAConnection-method}},
\code{\link{dbSendQuery,EXAConnection,character-method}},
\code{\link{dbWriteTable,EXAConnection,character,data.frame-method}}
}
\concept{EXAConnection related objects}
\concept{EXADriver related objects}
