% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXADBI-connection.R
\name{dbCurrentSchema}
\alias{dbCurrentSchema}
\title{Fetches and outputs the current schema from an EXASOL DB. Also updates EXAConnection metadata.}
\usage{
dbCurrentSchema(con, setSchema = NULL)
}
\arguments{
\item{con}{a valid EXAConnection}

\item{setSchema}{If not NULL, this schema will be opened and attached to EXAConnection metadata.}
}
\value{
an updated EXAConnection
}
\description{
Fetches and outputs the current schema from an EXASOL DB. Also updates EXAConnection metadata.
}
\seealso{
Other EXAConnection related objects: 
\code{\link{EXAConnection-class}},
\code{\link{dbBegin,EXAConnection-method}},
\code{\link{dbCommit,EXAConnection-method}},
\code{\link{dbConnect,EXADriver-method}},
\code{\link{dbEnd}()},
\code{\link{dbExistsTable,EXAConnection,character-method}},
\code{\link{dbGetQuery,EXAConnection,character-method}},
\code{\link{dbListFields,EXAConnection,character-method}},
\code{\link{dbListTables,EXAConnection-method}},
\code{\link{dbReadTable,EXAConnection,character-method}},
\code{\link{dbRemoveTable,EXAConnection,Id-method}},
\code{\link{dbRollback,EXAConnection-method}},
\code{\link{dbSendQuery,EXAConnection,character-method}},
\code{\link{dbWriteTable,EXAConnection,character,data.frame-method}}
}
\concept{EXAConnection related objects}
