% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXADBI.R
\name{dbCurrentSchema}
\alias{dbCurrentSchema}
\title{Fetches and outputs the current schema from an EXASOL DB. Also updates EXAConnection metadata.}
\usage{
dbCurrentSchema(con, setSchema = NULL)
}
\arguments{
\item{con}{a valid EXAConnection}
}
\value{
an updated EXAConnection
}
\description{
Fetches and outputs the current schema from an EXASOL DB. Also updates EXAConnection metadata.
}
\seealso{
Other EXAConnection related objects: 
\code{\link{EXAConnection-class}},
\code{\link{dbBegin}()},
\code{\link{dbCommit}()},
\code{\link{dbConnect}()},
\code{\link{dbEnd}()},
\code{\link{dbExistsTable}()},
\code{\link{dbGetQuery}()},
\code{\link{dbListFields}()},
\code{\link{dbListTables}()},
\code{\link{dbReadTable}()},
\code{\link{dbRemoveTable}()},
\code{\link{dbRollback}()},
\code{\link{dbSendQuery}()},
\code{\link{dbWriteTable}()}
}
\concept{EXAConnection related objects}
