% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXADBI-datatype.R
\name{dbDataType,EXAObject,ANY-method}
\alias{dbDataType,EXAObject,ANY-method}
\alias{dbDataType,EXAObject-method}
\title{dbDataType}
\usage{
\S4method{dbDataType}{EXAObject,ANY}(dbObj, obj, ...)
}
\arguments{
\item{dbObj}{A object inheriting from [EXADriver-class] or [EXAConnection-class].}

\item{obj}{An R object whose SQL type we want to determine.}

\item{...}{Unused.}
}
\description{
dbDataType
}
\section{Methods (by class)}{
\itemize{
\item \code{dbObj = EXAObject,obj = ANY}: Determine the EXASOL data type of an object.
}}

\seealso{
\code{\link{dbDataType,DBIObject-method}}
}
\author{
EXASOL AG <opensource@exasol.com>
}
