% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXADBI.R
\docType{methods}
\name{dbDataType}
\alias{dbDataType}
\title{Determine the EXASOL data type of an object.}
\usage{
\S4method{dbDataType}{EXAObject,ANY}(dbObj, obj, ...)
}
\description{
Determine the EXASOL data type of an object.
}
\seealso{
\code{\link{dbDataType,DBIObject-method}}
}

