% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXADBI-connection.R
\name{dbDisconnect,EXAConnection-method}
\alias{dbDisconnect,EXAConnection-method}
\title{dbDisconnect}
\usage{
\S4method{dbDisconnect}{EXAConnection}(conn)
}
\arguments{
\item{conn}{An EXAConnection object.}
}
\value{
A logical indicating success.
}
\description{
dbDisconnect
}
\section{Methods (by class)}{
\itemize{
\item \code{EXAConnection}: Disconnects the connection.
}}

\seealso{
\code{\link[DBI:dbDisconnect]{DBI::dbDisconnect()}}
}
