% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXADBI.R
\docType{methods}
\name{dbDisconnect}
\alias{dbDisconnect}
\title{Disconnects the connection.}
\usage{
\S4method{dbDisconnect}{EXAConnection}(conn)
}
\arguments{
\item{conn}{An EXAConnection object.}
}
\value{
A logical indicating success.
}
\description{
Disconnects the connection.
}

