% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXADBI.R
\name{dbEnd}
\alias{dbEnd}
\title{Ends a DB transaction. In EXASOL, it commits and reinstates the connection's standard autocommit mode.
This is an EXASOL specific addition to the DBI interface and may not work with other RDBMS.}
\usage{
dbEnd(conn, ...)
}
\arguments{
\item{conn}{An EXAConnection object}

\item{commit}{Logical. on TRUE the transaction is commmitted, otherwise rolled back.}
}
\value{
a logical indicating success.
}
\description{
Ends a DB transaction. In EXASOL, it commits and reinstates the connection's standard autocommit mode.
This is an EXASOL specific addition to the DBI interface and may not work with other RDBMS.
}
\seealso{
Other EXAConnection related objects: \code{\link{EXAConnection-class}},
  \code{\link{dbBegin}}, \code{\link{dbCommit}},
  \code{\link{dbConnect}}, \code{\link{dbCurrentSchema}},
  \code{\link{dbExistsTable}}, \code{\link{dbGetQuery}},
  \code{\link{dbReadTable}}, \code{\link{dbRemoveTable}},
  \code{\link{dbRollback}}, \code{\link{dbSendQuery}},
  \code{\link{dbWriteTable}}

Other transaction management functions: \code{\link{dbBegin}},
  \code{\link{dbCommit}}, \code{\link{dbRollback}}
}

