% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXADBI-transaction.R
\name{dbEnd}
\alias{dbEnd}
\alias{dbEnd,EXAConnection-method}
\title{dbEnd
Ends a DB transaction. In EXASOL, it commits and reinstates the connection's standard autocommit mode.
This is an EXASOL specific addition to the DBI interface and may not work with other RDBMS.}
\usage{
dbEnd(conn, ...)

\S4method{dbEnd}{EXAConnection}(conn, commit = TRUE, silent = FALSE)
}
\arguments{
\item{conn}{An EXAConnection object}

\item{...}{Additional parameters}

\item{commit}{Logical. on TRUE the transaction is commmitted, otherwise rolled back.}

\item{silent}{Logical. on FALSE, it shows logs in db actions.}
}
\value{
a logical indicating success.

a logical indicating success.
}
\description{
dbEnd
Ends a DB transaction. In EXASOL, it commits and reinstates the connection's standard autocommit mode.
This is an EXASOL specific addition to the DBI interface and may not work with other RDBMS.
}
\section{Methods (by class)}{
\itemize{
\item \code{EXAConnection}: Ends a DB transaction.
}}

\seealso{
Other EXAConnection related objects: 
\code{\link{EXAConnection-class}},
\code{\link{dbBegin,EXAConnection-method}},
\code{\link{dbCommit,EXAConnection-method}},
\code{\link{dbConnect,EXADriver-method}},
\code{\link{dbCurrentSchema}()},
\code{\link{dbExistsTable,EXAConnection,character-method}},
\code{\link{dbGetQuery,EXAConnection,character-method}},
\code{\link{dbListFields,EXAConnection,character-method}},
\code{\link{dbListTables,EXAConnection-method}},
\code{\link{dbReadTable,EXAConnection,character-method}},
\code{\link{dbRemoveTable,EXAConnection,character-method}},
\code{\link{dbRollback,EXAConnection-method}},
\code{\link{dbSendQuery,EXAConnection,character-method}},
\code{\link{dbWriteTable,EXAConnection,character,data.frame-method}}

Other transaction management functions: 
\code{\link{dbBegin,EXAConnection-method}},
\code{\link{dbCommit,EXAConnection-method}},
\code{\link{dbRollback,EXAConnection-method}}

Other EXADriver related objects: 
\code{\link{EXADriver-class}},
\code{\link{dbConnect,EXADriver-method}},
\code{\link{exasol_driver}()}

Other EXAConnection related objects: 
\code{\link{EXAConnection-class}},
\code{\link{dbBegin,EXAConnection-method}},
\code{\link{dbCommit,EXAConnection-method}},
\code{\link{dbConnect,EXADriver-method}},
\code{\link{dbCurrentSchema}()},
\code{\link{dbExistsTable,EXAConnection,character-method}},
\code{\link{dbGetQuery,EXAConnection,character-method}},
\code{\link{dbListFields,EXAConnection,character-method}},
\code{\link{dbListTables,EXAConnection-method}},
\code{\link{dbReadTable,EXAConnection,character-method}},
\code{\link{dbRemoveTable,EXAConnection,character-method}},
\code{\link{dbRollback,EXAConnection-method}},
\code{\link{dbSendQuery,EXAConnection,character-method}},
\code{\link{dbWriteTable,EXAConnection,character,data.frame-method}}
}
\concept{EXAConnection related objects}
\concept{EXADriver related objects}
\concept{None-DBI-API}
\concept{transaction management functions}
