% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXADBI.R
\docType{methods}
\name{dbExistsTable}
\alias{dbExistsTable}
\title{Checks if a table exists in an EXASOL DB.}
\usage{
\S4method{dbExistsTable}{EXAConnection,character}(conn, name, schema = "")
}
\arguments{
\item{conn}{An EXAConnection object.}

\item{name}{A fully qualified table name in the form schema.table.}

\item{schema}{Alternatively to `name`, a schema can be specified separately.}
}
\value{
A logical indicating if the table exists.
}
\description{
Checks if a table exists in an EXASOL DB.
}
\seealso{
Other DQL functions: \code{\link{dbClearResult}},
  \code{\link{dbFetch}}, \code{\link{dbGetQuery}},
  \code{\link{dbReadTable}}, \code{\link{dbSendQuery}}

Other EXAConnection related objects: \code{\link{EXAConnection-class}},
  \code{\link{dbBegin}}, \code{\link{dbCommit}},
  \code{\link{dbConnect}}, \code{\link{dbCurrentSchema}},
  \code{\link{dbEnd}}, \code{\link{dbGetQuery}},
  \code{\link{dbReadTable}}, \code{\link{dbRemoveTable}},
  \code{\link{dbRollback}}, \code{\link{dbSendQuery}},
  \code{\link{dbWriteTable}}
}

