% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXADBI-info.R
\name{dbGetInfo,EXAObject-method}
\alias{dbGetInfo,EXAObject-method}
\alias{dbGetInfo,EXADriver-method}
\alias{dbGetInfo,EXAConnection-method}
\alias{dbGetInfo,EXAResult-method}
\title{dbGetInfo}
\usage{
\S4method{dbGetInfo}{EXAObject}(dbObj)

\S4method{dbGetInfo}{EXADriver}(dbObj)

\S4method{dbGetInfo}{EXAConnection}(dbObj)

\S4method{dbGetInfo}{EXAResult}(dbObj)
}
\arguments{
\item{dbObj}{Existing EXAResult object.}
}
\value{
A named list.
}
\description{
dbGetInfo
}
\section{Methods (by class)}{
\itemize{
\item \code{EXAObject}: Returns metadata on a given EXAObject.

\item \code{EXADriver}: Returns metadata on a given EXADriver.

\item \code{EXAConnection}: Returns metadata on a given EXAConnection.

\item \code{EXAResult}: Returns metadata on a given EXAResult.
}}

\seealso{
\code{\link[DBI:dbGetInfo]{DBI::dbGetInfo()}}
}
\author{
EXASOL AG <opensource@exasol.com>

EXASOL AG <opensource@exasol.com>

EXASOL AG <opensource@exasol.com>

EXASOL AG <opensource@exasol.com>
}
