% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXADBI.R
\docType{methods}
\name{dbGetInfo}
\alias{dbGetInfo}
\title{Returns metadata on a given EXAObject.}
\usage{
\S4method{dbGetInfo}{EXAObject}(dbObj)
}
\arguments{
\item{dbObj}{An EXAObject.}
}
\value{
A named list.
}
\description{
Returns metadata on a given EXAObject.
}

