% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXADBI.R
\docType{methods}
\name{dbGetQuery}
\alias{dbGetQuery}
\title{Executes the query, fetches and returns the entire result set.}
\usage{
\S4method{dbGetQuery}{EXAConnection,character}(conn, statement, ...)
}
\arguments{
\item{conn}{An EXAConnection object.}

\item{statement}{An SQL query statement to be executed in an EXASOL DB.}

\item{...}{further arguements to be passed on to exa.readData.}
}
\value{
The result exa.readData, by default a data.frame containing the result set.
}
\description{
Executes the query, fetches and returns the entire result set.
}
\seealso{
Other DQL functions: \code{\link{dbClearResult}},
  \code{\link{dbExistsTable}}, \code{\link{dbFetch}},
  \code{\link{dbReadTable}}, \code{\link{dbSendQuery}}

Other EXAConnection related objects: \code{\link{EXAConnection-class}},
  \code{\link{dbBegin}}, \code{\link{dbCommit}},
  \code{\link{dbConnect}}, \code{\link{dbCurrentSchema}},
  \code{\link{dbEnd}}, \code{\link{dbExistsTable}},
  \code{\link{dbReadTable}}, \code{\link{dbRemoveTable}},
  \code{\link{dbRollback}}, \code{\link{dbSendQuery}},
  \code{\link{dbWriteTable}}
}

