% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXADBI-query.R
\name{dbGetQuery,EXAConnection,character-method}
\alias{dbGetQuery,EXAConnection,character-method}
\title{dbGetQuery}
\usage{
\S4method{dbGetQuery}{EXAConnection,character}(conn, statement, ...)
}
\arguments{
\item{conn}{An EXAConnection object.}

\item{statement}{An SQL query statement to be executed in an EXASOL DB.}

\item{...}{further arguments to be passed on to exa.readData.}
}
\value{
The result exa.readData, by default a data.frame containing the result set.
}
\description{
dbGetQuery
}
\section{Methods (by class)}{
\itemize{
\item \code{conn = EXAConnection,statement = character}: Executes the query, fetches and returns the entire result set.
}}

\seealso{
\code{\link[DBI:dbGetQuery]{DBI::dbGetQuery()}}

Other EXAConnection related objects: 
\code{\link{EXAConnection-class}},
\code{\link{dbBegin,EXAConnection-method}},
\code{\link{dbCommit,EXAConnection-method}},
\code{\link{dbConnect,EXADriver-method}},
\code{\link{dbCurrentSchema}()},
\code{\link{dbEnd}()},
\code{\link{dbExistsTable,EXAConnection,character-method}},
\code{\link{dbListFields,EXAConnection,character-method}},
\code{\link{dbListTables,EXAConnection-method}},
\code{\link{dbReadTable,EXAConnection,character-method}},
\code{\link{dbRemoveTable,EXAConnection,character-method}},
\code{\link{dbRollback,EXAConnection-method}},
\code{\link{dbSendQuery,EXAConnection,character-method}},
\code{\link{dbWriteTable,EXAConnection,character,data.frame-method}}

Other DQL functions: 
\code{\link{dbClearResult,EXAResult-method}},
\code{\link{dbExistsTable,EXAConnection,character-method}},
\code{\link{dbListFields,EXAConnection,character-method}},
\code{\link{dbListTables,EXAConnection-method}},
\code{\link{dbReadTable,EXAConnection,character-method}},
\code{\link{dbSendQuery,EXAConnection,character-method}},
\code{\link{fetch,EXAResult,numeric-method}}
}
\concept{DQL functions}
\concept{EXAConnection related objects}
