% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXADBI-result.R
\name{dbGetRowsAffected,EXAResult-method}
\alias{dbGetRowsAffected,EXAResult-method}
\title{dbGetRowsAffected}
\usage{
\S4method{dbGetRowsAffected}{EXAResult}(res)
}
\arguments{
\item{res}{Result from query.}
}
\description{
dbGetRowsAffected
}
\section{Methods (by class)}{
\itemize{
\item \code{EXAResult}: Get rows affected in result.
}}

\seealso{
\code{\link[DBI:dbGetRowsAffected]{DBI::dbGetRowsAffected()}}
}
