% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXADBI-object.R
\name{dbIsValid,EXAObject-method}
\alias{dbIsValid,EXAObject-method}
\title{dbIsValid}
\usage{
\S4method{dbIsValid}{EXAObject}(dbObj)
}
\arguments{
\item{dbObj}{An object that inherits EXAObject.}
}
\value{
A logical indicating if the connection still works.
}
\description{
dbIsValid
}
\section{Methods (by class)}{
\itemize{
\item \code{EXAObject}: Checks if an EXAObject is still valid.
}}

\seealso{
\code{\link[DBI:dbIsValid]{DBI::dbIsValid()}}
}
