% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXADBI.R
\name{dbIsValid}
\alias{dbIsValid}
\title{Checks if an EXAObject is still valid.}
\usage{
\S4method{dbIsValid}{EXAObject}(dbObj)
}
\arguments{
\item{conn}{An object that inherits EXAObject.}
}
\value{
A logical indicating if the connection still works.
}
\description{
Checks if an EXAObject is still valid.
}
