% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXADBI.R
\docType{methods}
\name{dbListTables}
\alias{dbListTables}
\title{Lists all tables in the DB.}
\usage{
\S4method{dbListTables}{EXAConnection}(conn, schema, ...)
}
\arguments{
\item{conn}{An EXAConnection object}

\item{schema}{Filter on DB schema}

\item{...}{further parameters passed on to `exa.readData()`}
}
\value{
a character vector. If no tables present, a character vector of length 0.
}
\description{
Lists all tables in the DB.
}
\seealso{
Other DQL functions: \code{\link{dbClearResult}},
  \code{\link{dbExistsTable}}, \code{\link{dbFetch}},
  \code{\link{dbGetQuery}}, \code{\link{dbListFields}},
  \code{\link{dbReadTable}}, \code{\link{dbSendQuery}}

Other EXAConnection related objects: \code{\link{EXAConnection-class}},
  \code{\link{dbBegin}}, \code{\link{dbCommit}},
  \code{\link{dbConnect}}, \code{\link{dbCurrentSchema}},
  \code{\link{dbEnd}}, \code{\link{dbExistsTable}},
  \code{\link{dbGetQuery}}, \code{\link{dbListFields}},
  \code{\link{dbReadTable}}, \code{\link{dbRemoveTable}},
  \code{\link{dbRollback}}, \code{\link{dbSendQuery}},
  \code{\link{dbWriteTable}}
}

