% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXADBI-table.R
\name{dbListTables,EXAConnection-method}
\alias{dbListTables,EXAConnection-method}
\title{dbListTables}
\usage{
\S4method{dbListTables}{EXAConnection}(conn, schema, ...)
}
\arguments{
\item{conn}{An EXAConnection object}

\item{schema}{Filter on DB schema}

\item{...}{further parameters passed on to `exa.readData()`}
}
\value{
a character vector. If no tables present, a character vector of length 0.
}
\description{
dbListTables
}
\section{Methods (by class)}{
\itemize{
\item \code{EXAConnection}: Lists all tables in the DB.
}}

\seealso{
\code{\link[DBI:dbListTables]{DBI::dbListTables()}}

Other EXAConnection related objects: 
\code{\link{EXAConnection-class}},
\code{\link{dbBegin,EXAConnection-method}},
\code{\link{dbCommit,EXAConnection-method}},
\code{\link{dbConnect,EXADriver-method}},
\code{\link{dbCurrentSchema}()},
\code{\link{dbEnd}()},
\code{\link{dbExistsTable,EXAConnection,character-method}},
\code{\link{dbGetQuery,EXAConnection,character-method}},
\code{\link{dbListFields,EXAConnection,character-method}},
\code{\link{dbReadTable,EXAConnection,character-method}},
\code{\link{dbRemoveTable,EXAConnection,Id-method}},
\code{\link{dbRollback,EXAConnection-method}},
\code{\link{dbSendQuery,EXAConnection,character-method}},
\code{\link{dbWriteTable,EXAConnection,character,data.frame-method}}

Other DQL functions: 
\code{\link{dbClearResult,EXAResult-method}},
\code{\link{dbExistsTable,EXAConnection,character-method}},
\code{\link{dbGetQuery,EXAConnection,character-method}},
\code{\link{dbListFields,EXAConnection,character-method}},
\code{\link{dbReadTable,EXAConnection,character-method}},
\code{\link{dbSendQuery,EXAConnection,character-method}},
\code{\link{fetch,EXAResult,numeric-method}}
}
\concept{DQL functions}
\concept{EXAConnection related objects}
