% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXADBI.R
\docType{methods}
\name{dbReadTable}
\alias{dbReadTable}
\title{Reads a DB table.}
\usage{
\S4method{dbReadTable}{EXAConnection,character}(conn, name, schema = "",
  order_col = NA, limit = NA, ...)
}
\arguments{
\item{conn}{An EXAConnection object.}

\item{name}{A fully qualified table name in the form schema.table.}

\item{schema}{Alternatively and with preference to `name`, a schema can be specified separately.}

\item{order_col}{A string containing columns to have the result set ordered upon, e.g. "col1 desc, col2"}

\item{limit}{A row limit to the result set.
CAUTION: a limit without order clause is non-deterministic in EXASOL,
i.e. subsequent runs of the same statement may deliver differing result sets.}

\item{...}{further arguements to be passed on to exa.readData.}
}
\value{
The result exa.readData, by default a data.frame containing the result set.
}
\description{
Reads a DB table.
}
\seealso{
Other DQL functions: \code{\link{dbClearResult}},
  \code{\link{dbExistsTable}}, \code{\link{dbFetch}},
  \code{\link{dbGetQuery}}, \code{\link{dbSendQuery}}

Other EXAConnection related objects: \code{\link{EXAConnection-class}},
  \code{\link{dbBegin}}, \code{\link{dbCommit}},
  \code{\link{dbConnect}}, \code{\link{dbCurrentSchema}},
  \code{\link{dbEnd}}, \code{\link{dbExistsTable}},
  \code{\link{dbGetQuery}}, \code{\link{dbRemoveTable}},
  \code{\link{dbRollback}}, \code{\link{dbSendQuery}},
  \code{\link{dbWriteTable}}
}

