% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXADBI-table.R
\name{dbRemoveTable,EXAConnection,Id-method}
\alias{dbRemoveTable,EXAConnection,Id-method}
\title{dbRemoveTable}
\usage{
\S4method{dbRemoveTable}{EXAConnection,Id}(conn, name, schema = "", cascade = FALSE)
}
\arguments{
\item{conn}{An EXAConnection object.}

\item{name}{a fully qualified table identifier (schema.table) or a simple table identifier, if the
parameter schema is specified.}

\item{schema}{An optional string specifying the DB schema. Caution: use either this parameter, OR give
a fully qualified identifier as giving both will result in an error.}

\item{cascade}{A logical indicating whether also foreign key constraints referencing the table to
be deleted shall be removed. Default is FALSE.}
}
\value{
A logicl indicating success.
}
\description{
dbRemoveTable
}
\section{Methods (by class)}{
\itemize{
\item \code{conn = EXAConnection,name = Id}: Removes a table.
}}

\seealso{
\code{\link[DBI:dbRemoveTable]{DBI::dbRemoveTable()}}

Other EXAConnection related objects: 
\code{\link{EXAConnection-class}},
\code{\link{dbBegin,EXAConnection-method}},
\code{\link{dbCommit,EXAConnection-method}},
\code{\link{dbConnect,EXADriver-method}},
\code{\link{dbCurrentSchema}()},
\code{\link{dbEnd}()},
\code{\link{dbExistsTable,EXAConnection,character-method}},
\code{\link{dbGetQuery,EXAConnection,character-method}},
\code{\link{dbListFields,EXAConnection,character-method}},
\code{\link{dbListTables,EXAConnection-method}},
\code{\link{dbReadTable,EXAConnection,character-method}},
\code{\link{dbRollback,EXAConnection-method}},
\code{\link{dbSendQuery,EXAConnection,character-method}},
\code{\link{dbWriteTable,EXAConnection,character,data.frame-method}}

Other DML functions: 
\code{\link{dbWriteTable,EXAConnection,character,data.frame-method}}
}
\concept{DML functions}
\concept{EXAConnection related objects}
