% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXADBI.R
\docType{methods}
\name{dbRemoveTable}
\alias{dbRemoveTable}
\title{Removes a table.}
\usage{
\S4method{dbRemoveTable}{EXAConnection}(conn, name, schema, cascade = FALSE)
}
\arguments{
\item{conn}{An EXAConnection object.}

\item{name}{a fully qualified table identifier (schema.table) or a simple table identifier, if the
parameter schema is specified.}

\item{schema}{An optional string specifying the DB schema. Caution: use either this parameter, OR give
a fully qualified identifier as giving both will result in an error.}

\item{cascade}{A logical indicating whether also foreign key constraints referencing the table to
be deleted shall be removed. Default is FALSE.}
}
\value{
A logicl indicating success.
}
\description{
Removes a table.
}
\seealso{
Other DML functions: \code{\link{dbWriteTable}}

Other EXAConnection related objects: \code{\link{EXAConnection-class}},
  \code{\link{dbBegin}}, \code{\link{dbCommit}},
  \code{\link{dbConnect}}, \code{\link{dbCurrentSchema}},
  \code{\link{dbEnd}}, \code{\link{dbExistsTable}},
  \code{\link{dbGetQuery}}, \code{\link{dbReadTable}},
  \code{\link{dbRollback}}, \code{\link{dbSendQuery}},
  \code{\link{dbWriteTable}}
}

