% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXADBI.R
\docType{methods}
\name{dbRollback}
\alias{dbRollback}
\title{Rolls the current DB transaction back.}
\usage{
\S4method{dbRollback}{EXAConnection}(conn)
}
\arguments{
\item{conn}{An EXAConnection object}
}
\value{
a logical indicating success.
}
\description{
Rolls the current DB transaction back.
}
\seealso{
Other EXAConnection related objects: \code{\link{EXAConnection-class}},
  \code{\link{dbBegin}}, \code{\link{dbCommit}},
  \code{\link{dbConnect}}, \code{\link{dbCurrentSchema}},
  \code{\link{dbEnd}}, \code{\link{dbExistsTable}},
  \code{\link{dbGetQuery}}, \code{\link{dbListFields}},
  \code{\link{dbListTables}}, \code{\link{dbReadTable}},
  \code{\link{dbRemoveTable}}, \code{\link{dbSendQuery}},
  \code{\link{dbWriteTable}}

Other transaction management functions: \code{\link{dbBegin}},
  \code{\link{dbCommit}}, \code{\link{dbEnd}}
}

