% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXADBI-query.R
\name{dbSendQuery,EXAConnection,character-method}
\alias{dbSendQuery,EXAConnection,character-method}
\title{dbSendQuery}
\usage{
\S4method{dbSendQuery}{EXAConnection,character}(
  conn,
  statement,
  schema = "",
  profile = TRUE,
  default_fetch_rec = 100,
  ...
)
}
\arguments{
\item{conn}{A valid EXAConnection}

\item{statement}{vector mode character : an SQL statement to be executed in EXASOL db}

\item{schema}{vector mode character : a focus schema. This must have write access for the result
set to be temporarily stored. If the user has only read permission on the schema to be read,
another schema may be entered here, and table identifiers in stmt parameter must be
fully qualified (schema.table).}

\item{profile}{logical, default TRUE : collect profiling information}

\item{default_fetch_rec}{numeric, default 100 :}

\item{...}{unused.}
}
\value{
EXAResult object which can be used for fetching rows. It also contains metadata.
}
\description{
dbSendQuery
}
\section{Methods (by class)}{
\itemize{
\item \code{conn = EXAConnection,statement = character}: Sends an SQL statment to an EXASOL DB, prepares for result fetching.
}}

\seealso{
\code{\link[DBI:dbSendQuery]{DBI::dbSendQuery()}}

Other EXAConnection related objects: 
\code{\link{EXAConnection-class}},
\code{\link{dbBegin,EXAConnection-method}},
\code{\link{dbCommit,EXAConnection-method}},
\code{\link{dbConnect,EXADriver-method}},
\code{\link{dbCurrentSchema}()},
\code{\link{dbEnd}()},
\code{\link{dbExistsTable,EXAConnection,character-method}},
\code{\link{dbGetQuery,EXAConnection,character-method}},
\code{\link{dbListFields,EXAConnection,character-method}},
\code{\link{dbListTables,EXAConnection-method}},
\code{\link{dbReadTable,EXAConnection,character-method}},
\code{\link{dbRemoveTable,EXAConnection,Id-method}},
\code{\link{dbRollback,EXAConnection-method}},
\code{\link{dbWriteTable,EXAConnection,character,data.frame-method}}

Other DQL functions: 
\code{\link{dbClearResult,EXAResult-method}},
\code{\link{dbExistsTable,EXAConnection,character-method}},
\code{\link{dbGetQuery,EXAConnection,character-method}},
\code{\link{dbListFields,EXAConnection,character-method}},
\code{\link{dbListTables,EXAConnection-method}},
\code{\link{dbReadTable,EXAConnection,character-method}},
\code{\link{fetch,EXAResult,numeric-method}}
}
\concept{DQL functions}
\concept{EXAConnection related objects}
