% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exasol.R
\docType{data}
\name{EMITS}
\alias{EMITS}
\title{EMITS output type of UDF script -- function emits any number of values.}
\format{\preformatted{ symbol EMITS
}}
\usage{
EMITS
}
\description{
EMITS output type of UDF script -- function emits any number of values.
}
\keyword{datasets}

