% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXADBI.R
\docType{class}
\name{EXAConnection-class}
\alias{EXAConnection-class}
\title{An Object holding a connection to an EXASOL Database.}
\description{
An Object holding a connection to an EXASOL Database.
}
\section{Slots}{

\describe{
\item{\code{init_connection_string}}{A string containing the ODBC connection sting used to initialise the connection.}

\item{\code{current_schema}}{A string reflecting the current schema.}

\item{\code{autocom_default}}{A logical indicating if autocommit is active.}

\item{\code{db_host}}{A string containing the hostname or IP.}

\item{\code{db_port}}{An integer containing the connection port.}

\item{\code{db_user}}{A string containing the DB user name.}

\item{\code{db_name}}{A string containing the database name.}

\item{\code{db_prod_name}}{A string containing the database product name.}

\item{\code{db_version}}{A string containing the database version.}

\item{\code{drv_name}}{A string containing the connection driver version.}

\item{\code{connection.string}}{RODBC}

\item{\code{handle_ptr}}{RODBC}

\item{\code{case}}{RODBC}

\item{\code{id}}{RODBC}

\item{\code{believedNRows}}{RODBC}

\item{\code{colQuote}}{RODBC}

\item{\code{tabQuote}}{RODBC}

\item{\code{encoding}}{RODBC}

\item{\code{rows_at_time}}{RODBC}
}}
\seealso{
\code{\link{DBIConnection-class}}

Other DBI classes: \code{\link{EXADriver-class}},
  \code{\link{EXAResult-class}}

Other EXAConnection related objects: \code{\link{dbBegin}},
  \code{\link{dbCommit}}, \code{\link{dbConnect}},
  \code{\link{dbCurrentSchema}}, \code{\link{dbEnd}},
  \code{\link{dbExistsTable}}, \code{\link{dbGetQuery}},
  \code{\link{dbReadTable}}, \code{\link{dbRemoveTable}},
  \code{\link{dbRollback}}, \code{\link{dbSendQuery}},
  \code{\link{dbWriteTable}}
}

