% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXADBI-datatype.R
\name{EXADataType}
\alias{EXADataType}
\alias{EXADataType,data.frame-method}
\alias{EXADataType,integer-method}
\alias{EXADataType,numeric-method}
\alias{EXADataType,logical-method}
\alias{EXADataType,Date-method}
\alias{EXADataType,POSIXct-method}
\alias{EXADataType,character-method}
\alias{EXADataType,factor-method}
\alias{EXADataType,list-method}
\alias{EXADataType,raw-method}
\alias{EXADataType,ANY-method}
\title{Generic for conversion to an EXASOL data type.}
\usage{
EXADataType(x)

\S4method{EXADataType}{data.frame}(x)

\S4method{EXADataType}{integer}(x)

\S4method{EXADataType}{numeric}(x)

\S4method{EXADataType}{logical}(x)

\S4method{EXADataType}{Date}(x)

\S4method{EXADataType}{POSIXct}(x)

\S4method{EXADataType}{character}(x)

\S4method{EXADataType}{factor}(x)

\S4method{EXADataType}{list}(x)

\S4method{EXADataType}{raw}(x)

\S4method{EXADataType}{ANY}(x)
}
\description{
Generic for conversion to an EXASOL data type.
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: Generic method for conversion of a dataframe to an EXASOL data type.

\item \code{integer}: Generic method for conversion of an integer to an EXASOL "INT".

\item \code{numeric}: Generic method for conversion of all other numerics to an EXASOL "DECIMAL(36,15)".

\item \code{logical}: Generic method for conversion of a logical to an EXASOL "BOOLEAN".

\item \code{Date}: Generic method for conversion of a Date to an EXASOL "DATE".

\item \code{POSIXct}: Generic method for conversion of a POSIXCT to an EXASOL "TIMESTAMP".

\item \code{character}: Generic method for conversion of a string to an EXASOL SQL string.

\item \code{factor}: Generic method for conversion of a factor to an EXASOL SQL string.

\item \code{list}: Generic method for conversion of a list to an EXASOL SQL string.

\item \code{raw}: Generic method for conversion of a "raw" to an EXASOL SQL string.

\item \code{ANY}: Generic method for conversion of "ANY" to an EXASOL SQL string.
}}

\keyword{internal}
