% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXADBI-driver.R
\docType{class}
\name{EXADriver-class}
\alias{EXADriver-class}
\title{An interface driver object to the EXASOL Database.}
\description{
An interface driver object to the EXASOL Database.
}
\section{Slots}{

\describe{
\item{\code{odbc_drv}}{A string containing the path to the EXASOL ODBC driver file.}
}}

\seealso{
\code{\link{DBIDriver-class}}

Other DBI classes: 
\code{\link{EXAConnection-class}},
\code{\link{EXAObject-class}},
\code{\link{EXAResult-class}}

Other EXADriver related objects: 
\code{\link{dbConnect,EXADriver-method}},
\code{\link{dbEnd}()},
\code{\link{exasol_driver}()}
}
\author{
EXASOL AG <opensource@exasol.com>
}
\concept{DBI classes}
\concept{EXADriver related objects}
