% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXADBI.R
\docType{class}
\name{EXADriver-class}
\alias{EXADriver-class}
\title{An interface driver object to the EXASOL Database.}
\description{
An interface driver object to the EXASOL Database.
}
\seealso{
\code{\link{DBIDriver-class}}

Other DBI classes: \code{\link{EXAConnection-class}},
  \code{\link{EXAResult-class}}

Other EXADriver related objects: \code{\link{dbConnect}}
}

