% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXADBI.R
\docType{class}
\name{EXAObject-class}
\alias{EXAObject-class}
\title{EXAObject class.}
\description{
The virtual object constituting a basis to all other EXA DBI Objects.
}
\seealso{
\code{\link{DBIObject-class}}

Other DBI classes: \code{\link{EXAConnection-class}},
  \code{\link{EXADriver-class}},
  \code{\link{EXAResult-class}}
}
