% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXADBI.R
\docType{class}
\name{EXAResult-class}
\alias{EXAResult}
\alias{EXAResult-class}
\title{An object that is associated with a result set in an EXASOL Database.}
\description{
An object that is associated with a result set in an EXASOL Database.
}
\section{Fields}{

\describe{
\item{\code{connection}}{An EXAConnection object.}

\item{\code{statement}}{A string containing the SQL query statement.}

\item{\code{rows_fetched}}{An int reflecting the rows already fetched from the DB.}

\item{\code{rows_affected}}{An int reflecting the length of the dataset in the DB.}

\item{\code{is_complete}}{A logical indcating if the query processing by the DB has been completed.}

\item{\code{with_output}}{A logical indicating whether the query produced a result set.}

\item{\code{profile}}{A data.frame containing profile information on the query.}

\item{\code{columns}}{A data.frame containing column metadata.}

\item{\code{temp_result_tbl}}{A string reflecting the name of the (temporary) table that holds the result set.}

\item{\code{query_sent_time}}{A POSIXct timestamp indicating when the query was sent to the DB.}

\item{\code{errors}}{A character vector containing errors.}

\item{\code{default_fetch_rec}}{An int reflecting the default fetch size.}
}}
\section{Methods}{

\describe{
\item{\code{addRowsFetched(x)}}{Add an int (the length of a newly fetched result set) to rows_fetched.}

\item{\code{close(commit = FALSE)}}{Frees up all resources, in particular drops the temporary table in the DB.}

\item{\code{refreshMetaData(x)}}{Refreshes the object's metadata.}
}}
\seealso{
Other DBI classes: \code{\link{EXAConnection-class}},
  \code{\link{EXADriver-class}}

Other EXAResult related objects: \code{\link{dbClearResult}},
  \code{\link{dbFetch}}
}

