% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exasol.R
\docType{package}
\name{exasol-package}
\alias{exasol}
\alias{exasol-package}
\title{EXASolution R Package}
\description{
The EXASolution R Package offers functionality to interact with
  the EXASolution database out of R programs. It is developed as a wrapper
  around ORDBC and extends ORDBC in two main aspects:

\enumerate{
  \item It offers fast data transfer between EXASolution and R, multiple
  times faster than RODBC. This is achieved by using a proprietary transfer
  channel which is optimized for batch loading.
  Please read the R help of \code{exa.readData()} and \code{exa.writeData()} for details.

  \item It makes it convenient to run parts of your R code in parallel on the
  EXASolution database, using EXASolution R UDF scripts behind the scenes.
  For example you can define an R function and execute it in parallel on
  different groups of data in an EXASolution table.
  Please read the R help of \code{exa.createScript()} function for details.
}

The help is available directly in R via:
\itemize{
  \item \code{help(exa.readData)}
  \item \code{help(exa.writeData)}
  \item \code{help(exa.createScript)}
}
}
\author{
EXASOL AG <support@exasol.com>
}
\keyword{distributed}
\keyword{in-memory}
\keyword{sql}

