% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXADBI-driver.R
\name{exasol_driver}
\alias{exasol_driver}
\alias{exa}
\title{Instantiates an EXADriver object.}
\usage{
exasol_driver(driver = NULL, silent = FALSE)

exa(driver = NULL, silent = FALSE)
}
\arguments{
\item{driver}{The path to an ODBC driver file. If missing, the driver installed on the system is used.
If "SYSTEM": the EXASOL ODBC driver installed on the system is used immediately.
Alternatively a path to an ODBC driver library can be provided.}

\item{silent}{If TRUE, no message is print.}
}
\value{
An EXADriver object.
}
\description{
Instantiates an EXADriver object.
}
\section{Functions}{
\itemize{
\item \code{exa}: An alias for \code{\link{exasol_driver}}
}}

\seealso{
Other EXADriver related objects: 
\code{\link{EXADriver-class}},
\code{\link{dbConnect,EXADriver-method}},
\code{\link{dbEnd}()}
}
\concept{EXADriver related objects}
