% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exa.log.R
\name{EXATrace}
\alias{EXATrace}
\title{EXATrace
Enables tracing for debugging purpose.}
\usage{
EXATrace(tracefile)
}
\arguments{
\item{tracefile}{The output tracefile}
}
\description{
EXATrace
Enables tracing for debugging purpose.
}
