% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXADBI-query.R
\name{fetch,EXAResult,numeric-method}
\alias{fetch,EXAResult,numeric-method}
\alias{fetch,EXAResult,missing-method}
\alias{dbFetch,EXAResult-method}
\title{fetch}
\usage{
\S4method{fetch}{EXAResult,numeric}(res, n = -1, ...)

\S4method{fetch}{EXAResult,missing}(res, n = -1, ...)

\S4method{dbFetch}{EXAResult}(res, n = -1, ...)
}
\arguments{
\item{res}{An EXAResult object.}

\item{n}{An int declaring the size of the subset to fetch. If missing, the whole subset is fetched.}

\item{...}{further arguments to be passed on to fetch.}
}
\description{
fetch

dbFetch
}
\section{Methods (by class)}{
\itemize{
\item \code{res = EXAResult,n = numeric}: Fetches a subset of an result set.

\item \code{res = EXAResult,n = missing}: Fetches a subset of an result set without indicated the size of the subset to fetch.

\item \code{EXAResult}: Fetches a subset of an result set.
}}

\seealso{
\code{\link[DBI:fetch]{DBI::fetch()}}

\code{\link[DBI:dbFetch]{DBI::dbFetch()}}

Other EXAResult related objects: 
\code{\link{EXAResult-class}},
\code{\link{dbClearResult,EXAResult-method}},
\code{\link{dbListFields,EXAConnection,character-method}}

Other DQL functions: 
\code{\link{dbClearResult,EXAResult-method}},
\code{\link{dbExistsTable,EXAConnection,character-method}},
\code{\link{dbGetQuery,EXAConnection,character-method}},
\code{\link{dbListFields,EXAConnection,character-method}},
\code{\link{dbListTables,EXAConnection-method}},
\code{\link{dbReadTable,EXAConnection,character-method}},
\code{\link{dbSendQuery,EXAConnection,character-method}}

Other EXAResult related objects: 
\code{\link{EXAResult-class}},
\code{\link{dbClearResult,EXAResult-method}},
\code{\link{dbListFields,EXAConnection,character-method}}

Other DQL functions: 
\code{\link{dbClearResult,EXAResult-method}},
\code{\link{dbExistsTable,EXAConnection,character-method}},
\code{\link{dbGetQuery,EXAConnection,character-method}},
\code{\link{dbListFields,EXAConnection,character-method}},
\code{\link{dbListTables,EXAConnection-method}},
\code{\link{dbReadTable,EXAConnection,character-method}},
\code{\link{dbSendQuery,EXAConnection,character-method}}

Other EXAResult related objects: 
\code{\link{EXAResult-class}},
\code{\link{dbClearResult,EXAResult-method}},
\code{\link{dbListFields,EXAConnection,character-method}}

Other DQL functions: 
\code{\link{dbClearResult,EXAResult-method}},
\code{\link{dbExistsTable,EXAConnection,character-method}},
\code{\link{dbGetQuery,EXAConnection,character-method}},
\code{\link{dbListFields,EXAConnection,character-method}},
\code{\link{dbListTables,EXAConnection-method}},
\code{\link{dbReadTable,EXAConnection,character-method}},
\code{\link{dbSendQuery,EXAConnection,character-method}}
}
\author{
EXASOL AG <opensource@exasol.com>

EXASOL AG <opensource@exasol.com>

EXASOL AG <opensource@exasol.com>
}
\concept{DQL functions}
\concept{EXAResult related objects}
