% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exasol.R
\docType{data}
\name{RETURNS}
\alias{RETURNS}
\title{RETURNS output type of UDF script -- function emits just a single value.}
\format{\preformatted{ symbol RETURNS
}}
\usage{
RETURNS
}
\description{
RETURNS output type of UDF script -- function emits just a single value.
}
\keyword{datasets}

