% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exasol.R
\docType{data}
\name{RETURNS}
\alias{RETURNS}
\title{RETURNS output type of UDF script -- function emits just a single value.}
\format{An object of class \code{character} of length 1.}
\usage{
RETURNS
}
\description{
RETURNS output type of UDF script -- function emits just a single value.
}
\keyword{datasets}

