% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exasol.R
\docType{data}
\name{SCALAR}
\alias{SCALAR}
\title{SCALAR input type of UDF script will call the function once for each record.}
\format{\preformatted{ symbol SCALAR
}}
\usage{
SCALAR
}
\description{
SCALAR input type of UDF script will call the function once for each record.
}
\keyword{datasets}

