% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exasol.R
\docType{data}
\name{SCALAR}
\alias{SCALAR}
\title{SCALAR input type of UDF script will call the function once for each record.}
\format{An object of class \code{character} of length 1.}
\usage{
SCALAR
}
\description{
SCALAR input type of UDF script will call the function once for each record.
}
\keyword{datasets}
